Update TxMA script and move files to common library
